requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
mysql-connector-python==9.5.0
requests==2.32.5
gunicorn==23.0.0
